Removed lock files & re-installed with NPM
App choked when trying to publish to expo.  I deleted these 2 lock
files and the node_modules folder, re-installed with "npm install",
and was able to publish to expo.

FWIW, I'm confused by switching between NPM and Yarn.  Is it bad to
delete the yarn lock file entirely? Is it ok to use two different
package managers?  Is that what caused the problem that required this
delete?  Unsure.  Committing anyway.