Work in Progress: create user
This is a stand-alone function that will successfully create a new
user account for the email address supplied in a POST body. You can
give it a try and see it working by using the firebase emulators
(see https://firebase.google.com/docs/rules/emulator-setup) and using
the cUrl sample in the function comments.

This functionality needs to be incorporated into another function that
has access to the parsed csv data so that users are created with the
right data.

(Creating a user probably won't be a stand-alone endpoint, unless there
is a compelling reason to do so. I just used a new function for testing.)